backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Api de site para casais"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "djangorestframework>=3.16.0",
    "psycopg2-binary>=2.9.10",
    "pillow>=11.2.1",
    "python-decouple>=3.8",
    "django-cors-headers>=4.7.0",
    "gunicorn>=23.0.0",
]

